// src/models/transaction-update.ts
export type TransactionUpdate = {
  id?: string;
  /**
   * Status of this transaction. You may set a transaction to `billed` or `canceled`. Billed
   * transactions cannot be changed.
   *
   * For manually-collected transactions, marking as `billed` is essentially issuing an invoice.
   */
  status?: TransactionStatus;
  /** Paddle ID of the customer that this transaction is for, prefixed with `ctm_`. */
  customerId?: string | null;
  /** Paddle ID of the address that this transaction is for, prefixed with `add_`. */
  addressId?: string | null;
  /** Paddle ID of the business that this transaction is for, prefixed with `biz_`. */
  businessId?: string | null;
  /** Your own structured key-value data. */
  customData?: Record<string, unknown> | null;
  /**
   * Supported three-letter ISO 4217 currency code. Must be `USD`, `EUR`, or `GBP` if
   * `collection_mode` is `manual`.
   */
  currencyCode?: CurrencyCode | null;
  origin?: PublicTransactionOrigin;
  /** Paddle ID of the subscription that this transaction is for, prefixed with `sub_`. */
  subscriptionId?: string | null;
  /**
   * Paddle ID of the invoice that this transaction is related to, prefixed with `inv_`. Used for
   * compatibility with the Paddle Invoice API, which is now deprecated. This field is scheduled to
   * be removed in the next version of the Paddle API.
   *
   * @deprecated
   */
  invoiceId?: string | null;
  /**
   * Invoice number for this transaction. Automatically generated by Paddle when you mark a
   * transaction as `billed` where `collection_mode` is `manual`.
   */
  invoiceNumber?: string | null;
  /**
   * How payment is collected for this transaction. `automatic` for checkout, `manual` for invoices.
   */
  collectionMode?: CollectionMode;
  /** Paddle ID of the discount to apply to this transaction, prefixed with `dsc_`. */
  discountId?: string | null;
  /** Apply a non-catalog discount to a transaction. Send one of `discount_id` or `discount`. */
  discount?: DiscountCustom | null;
  /** Details for invoicing. Required if `collection_mode` is `manual`. */
  billingDetails?: BillingDetailsUpdate | null;
  /**
   * Time period that this transaction is for. Set automatically by Paddle for subscription renewals
   * to describe the period that charges are for.
   */
  billingPeriod?: TimePeriod | null;
  /**
   * List of items on this transaction.
   *
   * When making a request, each object must contain either a `price_id` or a `price` object, and a
   * `quantity`.
   *
   * Include a `price_id` to charge for an existing catalog item, or a `price` object to charge for
   * a non-catalog item.
   */
  items?: TransactionItemUpdate[];
  details?: TransactionDetails;
  /**
   * List of payment attempts for this transaction, including successful payments. Sorted by
   * `created_at` in descending order, so most recent attempts are returned first.
   */
  payments?: TransactionPaymentAttempt[];
  /**
   * Paddle Checkout details for this transaction. You may pass a URL when creating or updating an
   * automatically-collected transaction, or when creating or updating a manually-collected
   * transaction where `billing_details.enable_checkout` is `true`.
   */
  checkout?: TransactionCheckoutCreate | null;
  createdAt?: Date;
  updatedAt?: Date;
  /**
   * RFC 3339 datetime string of when this transaction was marked as `billed`. `null` for
   * transactions that aren't `billed` or `completed`. Set automatically by Paddle.
   */
  billedAt?: Date | null;
};

export const transactionUpdateSchema: Schema<TransactionUpdate> = s.object<TransactionUpdate>({
  id: s.optional(s.string()),
  status: s.optional(s.lazy(() => transactionStatusSchema)),
  customerId: s.optionalNullable(s.string()),
  addressId: s.optionalNullable(s.string()),
  businessId: s.optionalNullable(s.string()),
  customData: s.optionalNullable(s.record(s.string(), s.unknown())),
  currencyCode: s.optionalNullable(s.lazy(() => currencyCodeSchema)),
  origin: s.optional(s.lazy(() => publicTransactionOriginSchema)),
  subscriptionId: s.optionalNullable(s.string()),
  invoiceId: s.optionalNullable(s.string()),
  invoiceNumber: s.optionalNullable(s.string()),
  collectionMode: s.optional(s.lazy(() => collectionModeSchema)),
  discountId: s.optionalNullable(s.string()),
  discount: s.optionalNullable(s.lazy(() => discountCustomSchema)),
  billingDetails: s.optionalNullable(s.lazy(() => billingDetailsUpdateSchema)),
  billingPeriod: s.optionalNullable(s.lazy(() => timePeriodSchema)),
  items: s.optional(s.array(s.lazy(() => transactionItemUpdateSchema))),
  details: s.optional(s.lazy(() => transactionDetailsSchema)),
  payments: s.optional(s.array(s.lazy(() => transactionPaymentAttemptSchema))),
  checkout: s.optionalNullable(s.lazy(() => transactionCheckoutCreateSchema)),
  createdAt: s.optional(s.dateTime()),
  updatedAt: s.optional(s.dateTime()),
  billedAt: s.optionalNullable(s.dateTime()),
  _keysMap: {
    customerId: "customer_id",
    addressId: "address_id",
    businessId: "business_id",
    customData: "custom_data",
    currencyCode: "currency_code",
    subscriptionId: "subscription_id",
    invoiceId: "invoice_id",
    invoiceNumber: "invoice_number",
    collectionMode: "collection_mode",
    discountId: "discount_id",
    billingDetails: "billing_details",
    billingPeriod: "billing_period",
    createdAt: "created_at",
    updatedAt: "updated_at",
    billedAt: "billed_at",
  },
});